Renaming to channel (singular) from provisioners(plural)
