+++ cmd/broker/ingress/main.go
@@ -31,7 +31,7 @@ import (
 	"k8s.io/client-go/kubernetes"
 	eventingv1alpha1 "knative.dev/eventing/pkg/apis/eventing/v1alpha1"
 	"knative.dev/eventing/pkg/broker/ingress"
-	"knative.dev/eventing/pkg/provisioners"
+	"knative.dev/eventing/pkg/channel"
 	"knative.dev/eventing/pkg/tracing"
 	"knative.dev/pkg/configmap"
 	"knative.dev/pkg/metrics"
@@ -50,8 +50,8 @@ type envConfig struct {
 }
 
 func main() {
-	logConfig := provisioners.NewLoggingConfig()
-	logger := provisioners.NewProvisionerLoggerFromConfig(logConfig).Desugar()
+	logConfig := channel.NewLoggingConfig()
+	logger := channel.NewProvisionerLoggerFromConfig(logConfig).Desugar()
 	defer flush(logger)
 	flag.Parse()
 	crlog.SetLogger(crlog.ZapLogger(false))
