+++ pkg/channel/message_receiver.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package provisioners
+package channel
 
 import (
 	"fmt"
