+++ pkg/channel/multichannelfanout/multi_channel_fanout_handler_test.go
@@ -26,7 +26,7 @@ import (
 	"github.com/google/go-cmp/cmp"
 	"go.uber.org/zap"
 	eventingduck "knative.dev/eventing/pkg/apis/duck/v1alpha1"
-	"knative.dev/eventing/pkg/provisioners/fanout"
+	"knative.dev/eventing/pkg/channel/fanout"
 )
 
 const (
