+++ pkg/channel/multichannelfanout/parse.go
