+++ pkg/reconciler/inmemorychannel/dispatcher/controller.go
@@ -21,7 +21,7 @@ import (
 	"time"
 
 	"knative.dev/eventing/pkg/inmemorychannel"
-	"knative.dev/eventing/pkg/provisioners/swappable"
+	"knative.dev/eventing/pkg/channel/swappable"
 	"knative.dev/eventing/pkg/reconciler"
 	"knative.dev/eventing/pkg/tracing"
